requirements.txt
asgiref==3.8.1
black==24.8.0
celery==5.3.6
dj-database-url==2.1.0
django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
gunicorn==21.2.0
isort==5.13.2
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
ruff==0.6.8
stripe==10.3.0
whitenoise==6.7.0
pytest==8.2.2
pytest-django==4.8.0

